CLI revamp plus new "show" feature
The previous default functionality has been moved to the
`hdrcopier copy` command.
There is a new command `hdrcopier show` which will print the HDR info to
the command line. This also has an option to print the ready-to-go CLI
args for supported programs, which currently are rav1e, x265, and
mkvmerge.